Refactoring to avoid call law level functions

+++ Shadows/jGame.cpp
@@ -250,7 +250,7 @@ void jGame::UpdateAppSetting()
 {
 	auto& appSetting =  jShadowAppSettingProperties::GetInstance();
 
-	appSetting.SpotLightDirection = Matrix::MakeRotateY(0.01).Transform(appSetting.SpotLightDirection);
+	appSetting.SpotLightDirection = Matrix::MakeRotateY(0.01f).Transform(appSetting.SpotLightDirection);
 
 	bool changedDirectionalLight = false;
 	if (appSetting.ShadowMapType == EShadowMapType::CSM_SSM)
@@ -562,7 +562,7 @@ void jGame::SpawnGraphTestFunc()
 	Vector OrthographicVector[90];
 	{
 		{
-			static jCamera* pCamera = jCamera::CreateCamera(Vector(0.0), Vector(0.0, 0.0, 1.0), Vector(0.0, 1.0, 0.0), DegreeToRadian(90), 10.0, 100.0, 100.0, 100.0, true);
+			static jCamera* pCamera = jCamera::CreateCamera(Vector(0.0f), Vector(0.0f, 0.0f, 1.0f), Vector(0.0f, 1.0f, 0.0f), DegreeToRadian(90.0f), 10.0f, 100.0f, 100.0f, 100.0f, true);
 			pCamera->UpdateCamera();
 			int cnt = 0;
 			auto MV = pCamera->Projection * pCamera->View;
@@ -572,10 +572,10 @@ void jGame::SpawnGraphTestFunc()
 			}
 
 			for (int i = 0; i < _countof(PerspectiveVector); ++i)
-				PerspectiveVector[i].z = (PerspectiveVector[i].z + 1.0) * 0.5;
+				PerspectiveVector[i].z = (PerspectiveVector[i].z + 1.0f) * 0.5f;
 		}
 		{
-			static jCamera* pCamera = jCamera::CreateCamera(Vector(0.0), Vector(0.0, 0.0, 1.0), Vector(0.0, 1.0, 0.0), DegreeToRadian(90), 10.0, 100.0, 100.0, 100.0, false);
+			static jCamera* pCamera = jCamera::CreateCamera(Vector(0.0f), Vector(0.0f, 0.0f, 1.0f), Vector(0.0f, 1.0f, 0.0f), DegreeToRadian(90.0f), 10.0f, 100.0f, 100.0f, 100.0f, false);
 			pCamera->UpdateCamera();
 			int cnt = 0;
 			auto MV = pCamera->Projection * pCamera->View;
@@ -585,7 +585,7 @@ void jGame::SpawnGraphTestFunc()
 			}
 
 			for (int i = 0; i < _countof(OrthographicVector); ++i)
-				OrthographicVector[i].z = (OrthographicVector[i].z + 1.0) * 0.5;
+				OrthographicVector[i].z = (OrthographicVector[i].z + 1.0f) * 0.5f;
 		}
 	}
 	std::vector<Vector2> graph1;
@@ -593,9 +593,9 @@ void jGame::SpawnGraphTestFunc()
 
 	float scale = 100.0f;
 	for (int i = 0; i < _countof(PerspectiveVector); ++i)
-		graph1.push_back(Vector2(i*2, PerspectiveVector[i].z * scale));
+		graph1.push_back(Vector2(static_cast<float>(i*2), PerspectiveVector[i].z * scale));
 	for (int i = 0; i < _countof(OrthographicVector); ++i)
-		graph2.push_back(Vector2(i*2, OrthographicVector[i].z* scale));
+		graph2.push_back(Vector2(static_cast<float>(i*2), OrthographicVector[i].z * scale));
 
 	auto graphObj1 = jPrimitiveUtil::CreateGraph2D({ 360, 350 }, {360, 300}, graph1);
 	jObject::AddUIDebugObject(graphObj1);
